+++ CHANGELOG.md
@@ -13,9 +13,12 @@
     - `pauseViewWithID:`
     - `pauseViewWithID:`
     - `stopAllViews:`
+- Added `enableAllConsents` initial config property to give all consents at init time
+- Added `giveAllConsents` method to give all consents
 - Added `enableAutomaticViewTracking` config for automatic track views
 - Added `automaticViewTrackingExclusionList` config for automatic view tracking exclusion list
 - Added `globalViewSegmentation` config to add set global view segmentation.
+- Deprecated `giveConsentForAllFeatures` method
 - Deprecated `CLYAutoViewTracking` in config
 - Deprecated existing view tracking methods and variables:  
     - `recordView:`